refactor: merge internal/timers.js and internal/timers.ts
