+++ node/internal/http.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 // Copyright Joyent and Node contributors. All rights reserved. MIT license.
 
-import { setUnrefTimeout } from "./timers.ts";
+import { setUnrefTimeout } from "../timers.ts";
 import { notImplemented } from "../_utils.ts";
 
 let utcCache: string | undefined;
