+++ node/internal/timers.js
@@ -2,11 +2,15 @@
 
 import { notImplemented } from "../_utils.ts";
 import { inspect } from "./util/inspect.js";
+import { validateNumber } from "./validators.js";
+import { ERR_OUT_OF_RANGE } from "../_errors.ts";
+import { emitWarning } from "../process.ts";
 
 // Timeout values > TIMEOUT_MAX are set to 1.
 export const TIMEOUT_MAX = 2 ** 31 - 1;
 
 export const kTimerId = Symbol("timerId");
+export const kTimeout = Symbol("timeout");
 const kRefed = Symbol("refed");
 
 // Timer constructor function.
@@ -53,4 +57,38 @@ Timeout.prototype.hasRef = function () {
 
 Timeout.prototype[Symbol.toPrimitive] = function () {
   return this[kTimerId];
+};
+
+/**
+ * @param {number} msecs
+ * @param {string} name
+ * @returns
+ */
+export function getTimerDuration(msecs, name) {
+  validateNumber(msecs, name);
+
+  if (msecs < 0 || !Number.isFinite(msecs)) {
+    throw new ERR_OUT_OF_RANGE(name, "a non-negative finite number", msecs);
+  }
+
+  // Ensure that msecs fits into signed int32
+  if (msecs > TIMEOUT_MAX) {
+    emitWarning(
+      `${msecs} does not fit into a 32-bit signed integer.` +
+        `\nTimer duration was truncated to ${TIMEOUT_MAX}.`,
+      "TimeoutOverflowWarning",
+    );
+
+    return TIMEOUT_MAX;
+  }
+
+  return msecs;
+}
+
+export default {
+  TIMEOUT_MAX,
+  kTimerId,
+  kTimeout,
+  Timeout,
+  getTimerDuration,
 };
