+++ node/module_all.ts
@@ -43,7 +43,7 @@ import internalStreamsAddAbortSignal from "./internal/streams/add-abort-signal.j
 import internalStreamsAddBufferList from "./internal/streams/buffer_list.js";
 import internalStreamsState from "./internal/streams/state.js";
 import internalTestBinding from "./internal/test/binding.ts";
-import internalTimers from "./internal/timers.ts";
+import internalTimers from "./internal/timers.js";
 import internalUtilInspect from "./internal/util/inspect.js";
 import net from "./net.ts";
 import os from "./os.ts";
